Drop untested graphic support from the build
Back in the day Tuletta (P8) machines would use Radeon GPU drivers. No
one has tested this in many moons. To save all systems are paying for
the kernel driver, disable it in the kernel configuration and drop the
firmware from the firenze, pseries and zz configurations.

A similar story applies for Matrox cards.

If someone comes hunting for support in the future they can revert this
commit.

Signed-off-by: Joel Stanley <joel@jms.id.au>